♻️(front) rename Heading.tsx file into index.tsx
We decided to rename every files containing a component into index.tsx
