+++ src/frontend/components/Headings/index.tsx
